refactor: `evm::debugger` -> `evm::debug`
