+++ evm/src/executor/mod.rs
@@ -25,7 +25,7 @@ pub use revm::SpecId;
 pub use revm::db::DatabaseRef;
 
 use self::inspector::InspectorStackConfig;
-use crate::{debugger::DebugArena, trace::CallTraceArena, CALLER};
+use crate::{debug::DebugArena, trace::CallTraceArena, CALLER};
 use bytes::Bytes;
 use ethers::{
     abi::{Abi, Detokenize, RawLog, Tokenize},
